apps/mip_ui_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
snowflake-connector-python>=3.6.0
python-dotenv>=1.0.0
